enshrine_shop/requirements.txt
asgiref==3.8.1
braintree==4.24.0
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
django-decouple==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Pillow==9.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
whitenoise==6.3.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1

